guacamole: remove common-ssh
