+++ guacamole/src/protocols/rdp/Makefile.am
@@ -142,7 +142,6 @@ endif
 libguac_client_rdp_la_CFLAGS = \
     -Werror -Wall -Iinclude    \
     @COMMON_INCLUDE@           \
-    @COMMON_SSH_INCLUDE@       \
     @LIBGUAC_INCLUDE@
 
 libguac_client_rdp_la_LDFLAGS = \
@@ -162,7 +161,6 @@ libguac_client_rdp_la_LIBADD =     \
 guacdr_cflags                = \
     -Werror -Wall -Iinclude    \
     @COMMON_INCLUDE@           \
-    @COMMON_SSH_INCLUDE@       \
     @LIBGUAC_INCLUDE@
 
 guacdr_ldflags =                   \
@@ -181,7 +179,6 @@ guacdr_libadd =     \
 guacai_cflags               =  \
     -Werror -Wall -Iinclude    \
     @COMMON_INCLUDE@           \
-    @COMMON_SSH_INCLUDE@       \
     @LIBGUAC_INCLUDE@
 
 guacai_ldflags =                   \
@@ -200,7 +197,6 @@ guacai_libadd =     \
 guacsnd_cflags               = \
     -Werror -Wall -Iinclude    \
     @COMMON_INCLUDE@           \
-    @COMMON_SSH_INCLUDE@       \
     @LIBGUAC_INCLUDE@
 
 guacsnd_ldflags =                  \
@@ -219,7 +215,6 @@ guacsnd_libadd =    \
 guacsvc_cflags               = \
     -Werror -Wall -Iinclude    \
     @COMMON_INCLUDE@           \
-    @COMMON_SSH_INCLUDE@       \
     @LIBGUAC_INCLUDE@
 
 guacsvc_ldflags =                  \
