+++ guacamole/src/protocols/ssh/_ssh.h
