+++ guacamole/src/protocols/ssh/dsa-compat.c
