+++ guacamole/src/protocols/ssh/dsa-compat.h
