+++ guacamole/src/protocols/ssh/key.h
