+++ guacamole/src/protocols/ssh/rsa-compat.c
