+++ guacamole/src/protocols/ssh/rsa-compat.h
